ref Изменил настройки подключения бд под использование SQlite ref Поправил все миграции всвязи с особенностями структуры SQlite

+++ database/migrations/2020_01_30_184723_add_column_off_method_view_items_table.php
@@ -18,14 +18,28 @@ class AddColumnOffMethodViewItemsTable extends Migration
                 if (Schema::hasColumn('view_items', 'id_method')) {
                     $table->renameColumn('id_method', 'on_method');
                 }
+            });
+        }
 
+        if (Schema::hasTable('view_items')) {
+            Schema::table('view_items', function (Blueprint $table) {
                 if (!Schema::hasColumn('view_items', 'off_method')) {
                     $table->unsignedInteger('off_method')->nullable()->after('id_method');
+                }
+            });
+        }
 
+        if (Schema::hasTable('view_items')) {
+            Schema::table('view_items', function (Blueprint $table) {
+                if (!Schema::hasColumn('view_items', 'off_method')) {
                     $table->foreign('off_method')->references('id')->on('methods')
                         ->onUpdate('cascade')->onDelete('set null');
                 }
+            });
+        }
 
+        if (Schema::hasTable('view_items')) {
+            Schema::table('view_items', function (Blueprint $table) {
                 if (!Schema::hasColumn('view_items', 'off_method_params')) {
                     $table->string('off_method_params', 100)->nullable();
                 }
